Move commands around, beg Pak for forgiveness ;-;
