+++ src/org/usfirst/frc/team694/robot/commands/auton/DrivetrainRotateDegreesPIDCommand.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team694.robot.commands;
+package org.usfirst.frc.team694.robot.commands.auton;
 
 import org.usfirst.frc.team694.robot.Robot;
