+++ src/org/usfirst/frc/team694/robot/commands/auton/SameSideRightAngleAutonCommand.java
@@ -3,7 +3,6 @@ package org.usfirst.frc.team694.robot.commands.auton;
 
 import org.usfirst.frc.team694.robot.FieldMapInterface;
 import org.usfirst.frc.team694.robot.Robot;
-import org.usfirst.frc.team694.robot.commands.DriveStraightWithRampingCommand;
 import org.usfirst.frc.team694.robot.commands.DrivetrainLineSensorReachNullCommand;
 import org.usfirst.frc.team694.robot.commands.DrivetrainRotateDegreesPIDCommand;
 import org.usfirst.frc.team694.robot.commands.GrabberOpenCommand;
